refactor(react): update structure of components
