--- src/modules/Home/components/Counter/index.ts
@@ -1,2 +1,0 @@
-export * from './Counter';
-export { default } from './Counter';
